+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Added an `auth` module to the `tower-reqwest` crate, which provides a way to
+  add an authorization header to the requests.
+
+- Added a `set-header` module to the `tower-reqwest` crate, which provides a way
+  to modify the request headers.
+
 ## [0.4.1] - 2024.12.04
 
 - Fix typos in the documentation.